Proper data exchange between pipeline processes and the supervisor
Added a custom multiprocessing manager that stores the raw filterbank data for reuse. Currently used by a single process, but eventually will be shared between the compute part of the pipeline and the plotting/archiving.
The underlying data table also allows us to read filterbank file only once, even if there are multiple candidates using it.

Improved the way that the data is read by the first module that needs it. Introduced a way to distinguish the modules and now it's the compute queue that manages which module has to read the data in instead of each module individually. Can remove a lot of repeat code.

Plotting and archiving moved to separate process. These modules will now wait for the message from the clustering supervisor. Not fully implemented yet though - currently just prints the received message.