Expanded the use of the manager(s)
Expanded the filterbank file data table. Now the table is checked for its size limits and the data is properly added and removed from the table.

Added extra candidate table which is used to store the candidate information for plotting and archiving modules. We now calculate a candidate 'hash' which is used as the key in the candidate dictionary 'table'. This information is also passed to the classifier, which later passes it back to the final modules.

Plotting and archiving modules now receive the data for processing. 

Current implementation is a bit broken in the sense that the filterbank file data used for plotting and archiving is the original data without any cleaning applied. This will have to be changed.