Safeguard for a missing run_summary.json
When discovering a new directory, run_summary.json might not yet exist. here we wait for it to appear for a minute before bailing out. Current implementation means the same directory will be retried every single time for a minute if something went wrong and this file simply doesn't exist. Will be changed later.