Lock the data table operations
There is a clear race condition that can result in a filterbank file being removed from the data table in one process between the check for its existence and eventual use by another process. This results in a KeyError exception as the key no longer exists. Currently using a very simple locking mechanism that locks the whole call to any changing function to matter the file.