Candidate queue with reprioritisation
Candidates now arranged in a priority in the candidate queue. Extended priority queue wrapped in multiprocessing manager to enable sharing between the processes.

On every get the queue checks whether the candidate is older than some predefined limit. This limit is chosen so that the candidate can still participate in the transient buffer clustering. If the candidate is older than that limit the queue is re-prioritised - older late candidates are moved towards the end of the queue. This can prevent us from losing too many candidates from TB clustering. Whatever is lost is lost and can be dealt with at a later time when we recover.

Currently ordering is done according to the priority number first, then MJD and then DM, all in increasing order.

Added code for CandidateQueue unit tests.