Fix(plotting): fixed valid sample check
Incorrect check in the kernel caused extra 16 samples to be processed and put on top of the existing data. If input data calculations were out of bounds, the code would quit with memory error. Otherwise first samples of the output would be corrupted.

Affected only the data which originally used odd number of blocks for processing.