Feat(watcher): new way of keeping track of processed files
We now keep an actual list (a set to be more precise) of detected and matched filterbank files. Removed reliance on the modification time of the file and the 'oldest' file to detect new filterbank files.
Previous method could occasionally cause issues when filterbank file is written before the .spccl file update and when multiple files are detected on the same iteration (happens around once in 5k candidates).