Fix(watcher): prevent NaNs from being sent to processing
We can catch .spccl file when it's still being written into and read an incomplete candidate like, especially when there are a lot of candidates. That causes NaNs to be sent through the pipeline causing errors in the candmaker module.

Here we wait for the .spccl file to be updated and not return any NaNs.