Fix(frbid): has include filterbank file name
Hash that includes just the DM, MJD and the absolute beam number causes problems if the candidate is present in the multiple files, resulting in the same hash. This in turn causes an extra candidate to be removed from the later file and the earlier file not to be removed from the data table at all.