Fix(filterbank): properly stop processing on deleted filterbank file
Data table now properly handles an exception when attempting to read a filterbank file that was deleted between being added to the processing queue and actually being processed. Pipeline is stopped immediately on detecting the missing file. 

Additionally properly handle a missing file exception on reading the data for the archive. This can happen if we exceed the size of the data table and have to reread the filterbank files.

Current implementation still goes through every candidate associated in that file.