Feat(modules): standardise the module abbreviation
Each transform module now has to have an abbreviation stored as a class variable. Module ID moved to a class variable as well. The list of abbreviations for started modules is now passed to the plotting configuration. All transform modules now included in on the plot (including the required modules) and in the order of processing.